modified temporary powers

+++ src/main/java/it/unibo/oop/bbgmm/Entity/Component/DamageComponent.java
@@ -7,7 +7,7 @@ public class DamageComponent extends AbstractEntityComponent implements Damage {
         this.damage = damage;
     }
     @Override
-    public double getDamage() {
+    public int getDamage() {
         return this.damage;
     }
